prevent deletion init
